pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiodns>=3.2.0",
    "aiohttp>=3.11.13",
    "async-timeout>=5.0.1",
    "colorama>=0.4.6",
    "dnspython>=2.7.0",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "trafilatura>=2.0.0",
]
